--- a/crypto/ec/ec_lib.c
+++ b/crypto/ec/ec_lib.c
@@ -114,6 +114,11 @@ void EC_pre_comp_free(EC_GROUP *group)
         EC_nistz256_pre_comp_free(group->pre_comp.nistz256);
 #endif
         break;
+    case PCT_sm2p256:
+#if !defined(OPENSSL_NO_SM2) && !defined(OPENSSL_NO_EC_SM2P_64_GCC_128)
+        EC_sm2p256_pre_comp_free(group->pre_comp.sm2p256);
+#endif
+        break;
 #ifndef OPENSSL_NO_EC_NISTP_64_GCC_128
     case PCT_nistp224:
         EC_nistp224_pre_comp_free(group->pre_comp.nistp224);
@@ -124,14 +129,10 @@ void EC_pre_comp_free(EC_GROUP *group)
     case PCT_nistp521:
         EC_nistp521_pre_comp_free(group->pre_comp.nistp521);
         break;
-    case PCT_sm2p256:
-        EC_sm2p256_pre_comp_free(group->pre_comp.sm2p256);
-        break;
 #else
     case PCT_nistp224:
     case PCT_nistp256:
     case PCT_nistp521:
-    case PCT_sm2p256:
         break;
 #endif
     case PCT_ec:
@@ -219,6 +220,11 @@ int EC_GROUP_copy(EC_GROUP *dest, const EC_GROUP *src)
         dest->pre_comp.nistz256 = EC_nistz256_pre_comp_dup(src->pre_comp.nistz256);
 #endif
         break;
+    case PCT_sm2p256:
+#if !defined(OPENSSL_NO_SM2) && !defined(OPENSSL_NO_EC_SM2P_64_GCC_128)
+        dest->pre_comp.sm2p256 = EC_sm2p256_pre_comp_dup(src->pre_comp.sm2p256);
+#endif
+        break;
 #ifndef OPENSSL_NO_EC_NISTP_64_GCC_128
     case PCT_nistp224:
         dest->pre_comp.nistp224 = EC_nistp224_pre_comp_dup(src->pre_comp.nistp224);
@@ -229,14 +235,10 @@ int EC_GROUP_copy(EC_GROUP *dest, const EC_GROUP *src)
     case PCT_nistp521:
         dest->pre_comp.nistp521 = EC_nistp521_pre_comp_dup(src->pre_comp.nistp521);
         break;
-    case PCT_sm2p256:
-        dest->pre_comp.sm2p256 = EC_sm2p256_pre_comp_dup(src->pre_comp.sm2p256);
-        break;
 #else
     case PCT_nistp224:
     case PCT_nistp256:
     case PCT_nistp521:
-    case PCT_sm2p256:
         break;
 #endif
     case PCT_ec: