@zzl360 2023-07-03
could you add which platform do you run command speed?