@FullyRobert 2023-07-05
This PR implements a relatively universal performance optimization that is applicable to most 64-bit platforms. Currently, it only requires little-endian byte order CPU architecture (a vast majority of common architectures, such as x86-64 and aarch64, are little-endian) and compiler support for 128-bit integers (native support has been implemented for most compilers except MSVC).